[TAN-5957] Extract `FilePlaceholder` to separate file
